+++ packages/author-head/__tests__/web/author-head.web.test.js
@@ -1,0 +1,8 @@
+/* eslint-env jest */
+
+import "jest-styled-components";
+import authorHeadTests from "../author-head.test.js";
+
+describe("AuthorHead test on web", () => {
+  authorHeadTests();
+});