+++ packages/card/card.stories.web.js
@@ -2,9 +2,9 @@
 import React from "react";
 import ReactDOMServer from "react-dom/server";
 import { storiesOf } from "dextrose/storiesOfOverloader";
+import { LateralSpacingDecorator } from "@times-components/storybook/decorators";
 import Card from "./card";
 import props from "./fixtures/card-props.json";
-import LateralSpacingDecorator from "../../storybook/decorators/lateral-spacing";
 
 props.date = new Date("2017-07-01T14:32:00.000Z");
 storiesOf("Card", module)