+++ packages/ad/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.5.4"></a>
+## [0.5.4](https://github.com/newsuk/times-components/compare/@times-components/ad@0.5.3...@times-components/ad@0.5.4) (2017-12-08)
+
+
+
+
+**Note:** Version bump only for package @times-components/ad
+
+<a name="0.5.3"></a>
+## [0.5.3](https://github.com/newsuk/times-components/compare/@times-components/ad@0.5.2...@times-components/ad@0.5.3) (2017-12-08)
+
+
+
+
+**Note:** Version bump only for package @times-components/ad
+
 <a name="0.5.2"></a>
 ## [0.5.2](https://github.com/newsuk/times-components/compare/@times-components/ad@0.5.0...@times-components/ad@0.5.2) (2017-12-04)