+++ packages/article/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.20.2"></a>
+## [0.20.2](https://github.com/newsuk/times-components/compare/@times-components/article@0.20.1...@times-components/article@0.20.2) (2017-12-08)
+
+
+
+
+**Note:** Version bump only for package @times-components/article
+
+<a name="0.20.1"></a>
+## [0.20.1](https://github.com/newsuk/times-components/compare/@times-components/article@0.20.0...@times-components/article@0.20.1) (2017-12-08)
+
+
+
+
+**Note:** Version bump only for package @times-components/article
+
 <a name="0.20.0"></a>
 # [0.20.0](https://github.com/newsuk/times-components/compare/@times-components/article@0.19.8...@times-components/article@0.20.0) (2017-12-07)