+++ packages/author-profile/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.30.1"></a>
+## [0.30.1](https://github.com/newsuk/times-components/compare/@times-components/author-profile@0.30.0...@times-components/author-profile@0.30.1) (2017-12-08)
+
+
+
+
+**Note:** Version bump only for package @times-components/author-profile
+
+<a name="0.30.0"></a>
+# [0.30.0](https://github.com/newsuk/times-components/compare/@times-components/author-profile@0.29.1...@times-components/author-profile@0.30.0) (2017-12-08)
+
+
+### Features
+
+* make author head responsive ([#434](https://github.com/newsuk/times-components/issues/434)) ([5a07f59](https://github.com/newsuk/times-components/commit/5a07f59))
+
+
+
+
+<a name="0.29.1"></a>
+## [0.29.1](https://github.com/newsuk/times-components/compare/@times-components/author-profile@0.29.0...@times-components/author-profile@0.29.1) (2017-12-08)
+
+
+
+
+**Note:** Version bump only for package @times-components/author-profile
+
 <a name="0.29.0"></a>
 # [0.29.0](https://github.com/newsuk/times-components/compare/@times-components/author-profile@0.28.2...@times-components/author-profile@0.29.0) (2017-12-07)