+++ packages/card/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.12.2"></a>
+## [0.12.2](https://github.com/newsuk/times-components/compare/@times-components/card@0.12.1...@times-components/card@0.12.2) (2017-12-08)
+
+
+
+
+**Note:** Version bump only for package @times-components/card
+
+<a name="0.12.1"></a>
+## [0.12.1](https://github.com/newsuk/times-components/compare/@times-components/card@0.12.0...@times-components/card@0.12.1) (2017-12-08)
+
+
+
+
+**Note:** Version bump only for package @times-components/card
+
 <a name="0.12.0"></a>
 # [0.12.0](https://github.com/newsuk/times-components/compare/@times-components/card@0.11.3...@times-components/card@0.12.0) (2017-12-07)